% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obif.ffa.R
\name{obif.ffa}
\alias{obif.ffa}
\title{Full factorial analysis}
\usage{
obif.ffa(dataAR, design, contrast)
}
\description{
Primary function to perform full factorial analysis in an Ananlysis-Ready dataset
}
\details{
@param dataAR A numeric data matrix with the analyis-ready feature expression values where row are features and columns are samples

 @param design A design matrix to be applied for expression analysis

 @param contrast A contrast matrix to be applied for contrast analysis

 @return A data frame with the individual p-values per feature for main effects, simple main effects and interaction effects.
}
